pyproject.toml
[tool.poetry]
name = "fasthtml-tailwind"
version = "0.1.0"
description = ""
authors = ["Jose Antonio Dominguez <chamot11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
